backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-docx>=0.8.11
reportlab>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# Authentication
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.28.0
aiofiles>=23.2.0
docxtpl>=0.16.0
docx2pdf>=0.1.8
openpyxl>=3.1.2
# Testing
pytest>=7.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
